+++ src/LibSSH.jl
@@ -9,7 +9,7 @@ using DocStringExtensions
 
 include("bindings.jl")
 using .lib
-import .lib: LibSSHException, ssh_options_get, ssh_options_set, SSH_OK, SSH_ERROR, SSH_AGAIN, SSH_EOF
+import .lib: LibSSHException, SSH_OK, SSH_ERROR, SSH_AGAIN, SSH_EOF
 
 
 """